+++ doc/CHANGELOG.md
@@ -3,6 +3,7 @@
 * `#1755` Migrate helper-tests for issues into specs for work package
 * `#1766` Fixed bug: Viewing diff of Work Package description results in error 500
 * `#1789` Move validation to Work Package
+* `#1793` Data Migration Journals
 * `#1800` Add settings to change software name and URL and add additional footer content
 * `#1808` Add option to log user for each request
 * `#1876` Timelines do not show work packages when there is no status reporting